wip: restore qlty smell threshold
